[wip] move ssd_button related things into own file
