+++ src/desktop.c
@@ -384,6 +384,7 @@ get_cursor_context(struct server *server)
 				return ret;
 			case LAB_NODE_DESC_NODE:
 			case LAB_NODE_DESC_TREE:
+			case LAB_NODE_DESC_SSD_ROUNDED:
 				break;
 			}
 		}
