+++ src/ssd/meson.build
@@ -1,8 +1,9 @@
 labwc_sources += files(
   'resize_indicator.c',
   'ssd.c',
+  'ssd_border.c',
+  'ssd_button.c',
+  'ssd_extents.c',
   'ssd_part.c',
   'ssd_titlebar.c',
-  'ssd_border.c',
-  'ssd_extents.c',
 )
